Correction from Shin-Liang
MF: Hi Shin-Liang,
Sorry for the late reply (I was outside UK for the bank holiday weekend).

On Fri, Aug 24, 2012 at 8:52 AM, S.L. Chin <slc60@cam.ac.uk> wrote:
SL: Hi Matteo,

Thanks for the reply. In your email below, I believe you meant to really say to use "a0_axis" for a change in the anisotropy axis? I

MF: Ops... Yes, that's true.

SL: managed to get this working with a syntax something like this:
   p = anisotropy.quantities["a0_axis"]
   p.set_value(Value(create_setter([1,0,0], [0,1,0]), SI(1)))

Also, I believe there is a slight error with the spatially.py code in line 43-46. Perhaps the create_setter function is better to have written as:

 value_middle = \
   ([0.5*(v + value_top[i]) for i, v in enumerate(value_bottom)]
     if hasattr(value_top, "__iter__")
     else 0.5*(value_top + value_bottom))


where the hassattr is evaluated upon value_top instead of value_mid as is described in the spatially.pdf? this is because I found that in the original the create_setter function worked very well for a constant, it did not do so for a vector.

MF: You are right, thanks for spotting this :-)


Finally, on a question related to the macrogeometry I calculated:

x_lattice=4.64
y_lattice=13.92
z_lattice=0.0
lattice_points=[]
for xi in range (-18,18):
        for yi in range (-6,6):
                 lattice_points.append([xi*x_lattice,yi*y_lattice,0.0*z_lattice]) pbc = nmag.SetLatticePoints(vectorlist=lattice_points, scalefactor=SI(1e-9,'m'))

The output of the calculations (as displayed using mayavi2) shows that there is 1x2 periodicity, i.e. the magnetization repeats for every two cells in the y-direction and every cell in the x-direction. I would like to know does this sounds right since the resultant macrogeometry is a perfect square, shouldn't it be a perfect repetittion at every cell?

MF: I am not sure I understand the problem. I can say for sure that what you see with MayaVi is just the primary cell, not the entire periodic system. In other words, Mayavi does not show the repeated magnetisation vectors.

Also, if you are simulating bulk material (or if - in your system - the primary cell touches any of the image cells) I would suggest to make the period slightly larger. For example, if you are simulating a cube 10 x 20 x 30, I would make sure the distances between the cube centres are 10.01 x 20.01 x 30.01. This is necessary to circumvent a bug in the calculation of the demagnetising field. Notice, that despite doing this you introduce a small gap, the result will be mostly unaffected (the exchange interaction works as before, keeping the magnetisation of neighbouring image cells aligned, while the surface charges between neighbouring cells cancel out appropriately). If your system does not involve "touching" image cells, then these considerations do not apply.
Best wishes,
Matteo